Part 1: start task4
